move fraction,numer,denom to core.mul
